pyproject.toml
[project]
name = "homen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3.0",
    "django-environ>=0.12.0",
    "psycopg[binary]>=3.2.0",
    "gunicorn>=23.0.0",
    "requests>=2.32.0",
    "cryptography>=43.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
    "factory-boy>=3.3.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "common", "config"]